server/requirements.txt
Django
psycopg2-binary>=2.8
djangorestframework
firebase-admin
requests==2.22.0
python-oauth2
oauthlib
oauth2
urllib3
#python-firebase
git+https://github.com/ozgur/python-firebase
markdown
django-filter
django-cors-headers
requests_oauthlib